refactor(esp-nvs): split code into more atomic modules
